feat(worldgen): extract inn/tavern placement into buildInnAndMarkTavern and wire into town_gen.js; update imports.
